got rid of injection attack risks

+++ server.js
@@ -142,8 +142,8 @@ app.post('/make_goal', async (req, res) => {
 });
 
 app.post('/edit_goal', async(req, res) => {
-    db.run(`UPDATE goals SET aim=?, end_date=? WHERE id=?`,
-         [req.body.numTimes, req.body.newEnd, req.body.id],
+    db.run(`UPDATE goals SET aim=?, end_date=? WHERE id=? AND user_id=?`,
+         [req.body.numTimes, req.body.newEnd, req.body.id, req.session.user],
          (err) => {
             if (err) return res.redirect('/settings?error=update');
             res.redirect('/settings?updated=1');
@@ -151,8 +151,8 @@ app.post('/edit_goal', async(req, res) => {
 });
 
 app.post('/delete_goal', async(req, res) => {
-    db.run(`DELETE FROM goals WHERE id=?`,
-         [req.body.id],
+    db.run(`DELETE FROM goals WHERE id=? AND user_id=?`,
+      [req.body.id, req.session.user],
          (err) => {
             if (err) return res.redirect('/settings?error=update');
             res.redirect('/settings?updated=1');
